hotfix: 신청서 결과 요약 조회 시 중복 응답 여부와 총 응답 수 반환하도록 수정

+++ src/main/java/net/causw/application/dto/form/response/QuestionSummaryResponseDto.java
@@ -31,7 +31,7 @@ public class QuestionSummaryResponseDto {
     private List<OptionSummaryResponseDto> optionSummarieList;
 
     @Schema(description = "총 응답 수", example = "10")
-    private Integer numOfReply;
+    private Long numOfReply;
 
     @Schema(description = "복수 정답 여부(객관식일 때만)", defaultValue = "false", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
     private Boolean isMultiple;
